MLBEDSW-9669: Improve BufferView reader/writer semantics
 - Update Buffer to use more standard size_t byte count
   for initialisation use case.
 - Separate BufferValues into two classes:
    - Implement BufferReader to provide polymorphic
      access and iterator support reading typed buffers.
    - Implement BufferWriter for direct-access read/writes.
    - Catch linear indexing if strides indicate non-linear
      buffer layout.
    - Provide shape-indexed overloads for coordinate lookup.
 - Separate concept of buffer size (which is in bytes) from
   element count. Some code misused this equivalence for
   8-bit types.
 - Require explicitly named access to the raw buffer data.
 - Remove two unused buffer slicing functions that did not
   honour strided buffer layout.
 - Added basic test harness for BufferView expected use case
   and fixed existing test case that uses the buffer view.
 - Fixed code using direct buffer access that assumed underlying
   data-type without checking first.
 - Allow the Shape dot-product to assume '1' axes on left
   operand only.
 - Fixed vc compiler type conversion issues with DMA offset
   type and quantisation scale maths.

+++ ethosu/regor/architecture/ethosu65/ethos_u65_register_cs_generator.cpp
@@ -154,14 +154,14 @@ void EthosU65RCSGenerator::GenerateDMA(const HLCDMA *dma, MemoryAccesses &memory
     if ( strideMode != dma_stride_mode::D1 )
     {
         Emit(isa::npu_set_dma0_size0_t(size0));
-        int skip0 = destStride0 - dma->length;
+        uint64_t skip0 = destStride0 - dma->length;
         Emit(isa::npu_set_dma0_skip0_t(skip0));
     }
 
     if ( strideMode == dma_stride_mode::D3 )
     {
         Emit(isa::npu_set_dma0_size1_t(size1));
-        int skip1 = destStride1 - (dma->length * size0);
+        uint64_t skip1 = destStride1 - (dma->length * size0);
         Emit(isa::npu_set_dma0_skip1_t(skip1));
     }
